Move code-hashing strategy to non-PoA location
