+++ src/Stratis.Bitcoin.Features.SmartContracts/Interfaces/IContractCodeHashingStrategy.cs
@@ -1,4 +1,4 @@
-namespace Stratis.Bitcoin.Features.SmartContracts.PoA
+namespace Stratis.Bitcoin.Features.SmartContracts.Interfaces
 {
     public interface IContractCodeHashingStrategy
     {
